backend/requirements.txt
Flask>=3.0.0
flask-cors>=4.0.0
pydantic>=2.5.0
openai>=1.0.0
python-dotenv>=1.0.0
supabase>=2.4.0
httpx>=0.27.0